Move ClassifyImageEvent from events to vision_backend.
As we go on to add more types of events, it seems more manageable for each subclass to live in the app that relates to the type of event it is.
